refactored and add test for update project file, use tempfile crate
