+++ src/lib/src/cli/mod.rs
@@ -1,16 +1,6 @@
-pub mod dialog;
-pub use dialog::*;
-pub mod install_node;
-pub use install_node::*;
-pub mod install_wallet;
-pub use install_wallet::*;
-pub mod uninstall;
-pub use uninstall::*;
+pub mod node;
+pub use node::*;
+pub mod wallet;
+pub use wallet::*;
 pub mod update;
 pub use update::*;
-pub mod sync;
-pub use sync::*;
-pub mod run;
-pub use run::*;
-pub mod setup;
-pub use setup::*;
