+++ src/lib/src/cli/node/uninstall.rs
@@ -12,12 +12,6 @@ pub async fn uninstall_cli() -> Result<()> {
     Ok(())
 }
 
-pub async fn uninstall_wallet() -> Result<()> {
-    log::info!("Uninstalling cardano-wallet");
-    log::warn!("Not yet implemented");
-    Ok(())
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
@@ -31,11 +25,6 @@ mod test {
     #[tokio::test]
     async fn test_uninstall_cli() {
         let result = uninstall_cli().await.is_ok();
-        assert!(result);
-    }
-    #[tokio::test]
-    async fn test_uninstall_wallet() {
-        let result = uninstall_wallet().await.is_ok();
         assert!(result);
     }
 }
