+++ src/lib/src/cli/wallet/install.rs
