+++ src/lib/src/cli/wallet/mod.rs
@@ -1,0 +1,6 @@
+pub mod install;
+pub use install::*;
+pub mod uninstall;
+pub use uninstall::*;
+pub mod setup;
+pub use setup::*;
